# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Test Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BloodTest {
  id: number;
  fileName: string;
  /** @nullable */
  testDate: string | null;
  /** @nullable */
  labName: string | null;
  /** @nullable */
  patientName: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  resultCount: number;
}

export interface BloodTestResult {
  id: number;
  bloodTestId: number;
  markerName: string;
  /** @nullable */
  value: number | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  referenceRangeLow: number | null;
  /** @nullable */
  referenceRangeHigh: number | null;
  /** @nullable */
  status: string | null;
  /** @nullable */
  rawText: string | null;
}

export interface BloodTestWithResults {
  id: number;
  fileName: string;
  /** @nullable */
  testDate: string | null;
  /** @nullable */
  labName: string | null;
  /** @nullable */
  patientName: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  results: BloodTestResult[];
}

export interface UploadBloodTestBody {
  fileName: string;
  pdfBase64: string;
  /** @nullable */
  notes?: string | null;
}

export interface BloodTestSummary {
  totalTests: number;
  totalMarkers: number;
  abnormalCount: number;
  /** @nullable */
  latestTestDate: string | null;
  commonMarkers: string[];
}

export interface MarkerHistory {
  markerName: string;
  /** @nullable */
  testDate: string | null;
  /** @nullable */
  value: number | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  status: string | null;
  /** @nullable */
  referenceRangeLow: number | null;
  /** @nullable */
  referenceRangeHigh: number | null;
  bloodTestId: number;
}

export interface ChatConversation {
  id: number;
  title: string;
  /** @nullable */
  bloodTestId: number | null;
  createdAt: string;
}

export interface ChatMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatConversationWithMessages {
  id: number;
  title: string;
  /** @nullable */
  bloodTestId: number | null;
  createdAt: string;
  messages: ChatMessage[];
}

export interface CreateChatConversationBody {
  title: string;
  /** @nullable */
  bloodTestId?: number | null;
}

export interface SendChatMessageBody {
  content: string;
}

export interface ApiError {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface SendOpenaiVoiceMessageBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export type ListMarkerHistoryParams = {
  markerName?: string;
};