# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Blood Test Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blood-tests:
    get:
      operationId: listBloodTests
      tags: [blood-tests]
      summary: List all blood tests
      responses:
        "200":
          description: List of blood tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BloodTest"
    post:
      operationId: uploadBloodTest
      tags: [blood-tests]
      summary: Upload a blood test PDF and extract results
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadBloodTestBody"
      responses:
        "201":
          description: Blood test created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodTestWithResults"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /blood-tests/{id}:
    get:
      operationId: getBloodTest
      tags: [blood-tests]
      summary: Get a blood test with all results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blood test with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodTestWithResults"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteBloodTest
      tags: [blood-tests]
      summary: Delete a blood test
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /blood-tests/markers:
    get:
      operationId: listMarkerHistory
      tags: [blood-tests]
      summary: Get history of a specific marker across all tests
      parameters:
        - name: markerName
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Marker history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarkerHistory"
  /chat/conversations:
    get:
      operationId: listChatConversations
      tags: [chat]
      summary: List all chat conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatConversation"
    post:
      operationId: createChatConversation
      tags: [chat]
      summary: Create a new chat conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatConversation"
  /chat/conversations/{id}:
    get:
      operationId: getChatConversation
      tags: [chat]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteChatConversation
      tags: [chat]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /chat/conversations/{id}/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get a streaming AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    BloodTest:
      type: object
      properties:
        id:
          type: integer
        fileName:
          type: string
        testDate:
          type: ["string", "null"]
        labName:
          type: ["string", "null"]
        patientName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        resultCount:
          type: integer
      required:
        - id
        - fileName
        - testDate
        - labName
        - patientName
        - notes
        - createdAt
        - resultCount
    BloodTestResult:
      type: object
      properties:
        id:
          type: integer
        bloodTestId:
          type: integer
        markerName:
          type: string
        value:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        referenceRangeLow:
          type: ["number", "null"]
        referenceRangeHigh:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
        rawText:
          type: ["string", "null"]
      required:
        - id
        - bloodTestId
        - markerName
        - value
        - unit
        - referenceRangeLow
        - referenceRangeHigh
        - status
        - rawText
    BloodTestWithResults:
      type: object
      properties:
        id:
          type: integer
        fileName:
          type: string
        testDate:
          type: ["string", "null"]
        labName:
          type: ["string", "null"]
        patientName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        results:
          type: array
          items:
            $ref: "#/components/schemas/BloodTestResult"
      required:
        - id
        - fileName
        - testDate
        - labName
        - patientName
        - notes
        - createdAt
        - results
    UploadBloodTestBody:
      type: object
      properties:
        fileName:
          type: string
        pdfBase64:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - fileName
        - pdfBase64
...
    MarkerHistory:
      type: object
      properties:
        markerName:
          type: string
        testDate:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        referenceRangeLow:
          type: ["number", "null"]
        referenceRangeHigh:
          type: ["number", "null"]
        bloodTestId:
          type: integer
      required:
        - markerName
        - testDate
        - value
        - unit
        - status
        - referenceRangeLow
        - referenceRangeHigh
        - bloodTestId
    ChatConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        bloodTestId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - bloodTestId
        - createdAt
    ChatConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        bloodTestId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required:
        - id
        - title
        - bloodTestId
        - createdAt
        - messages
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateChatConversationBody:
      type: object
      properties:
        title:
          type: string
        bloodTestId:
          type: ["integer", "null"]
      required:
        - title
    SendChatMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...