# lib/api-zod/src/generated/types/bloodTestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Test Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface BloodTestResult {
  id: number;
  bloodTestId: number;
  markerName: string;
  /** @nullable */
  value: number | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  referenceRangeLow: number | null;
  /** @nullable */
  referenceRangeHigh: number | null;
  /** @nullable */
  status: string | null;
  /** @nullable */
  rawText: string | null;
}