# lib/api-zod/src/generated/types/bloodTestWithResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Test Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { BloodTestResult } from "./bloodTestResult";

export interface BloodTestWithResults {
  id: number;
  fileName: string;
  /** @nullable */
  testDate: string | null;
  /** @nullable */
  labName: string | null;
  /** @nullable */
  patientName: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
  results: BloodTestResult[];
}