# lib/api-zod/src/generated/types/chatConversation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Test Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatConversation {
  id: number;
  title: string;
  /** @nullable */
  bloodTestId: number | null;
  createdAt: Date;
}