# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Test Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: Date;
}