# lib/api-zod/src/generated/types/createChatConversationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Test Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateChatConversationBody {
  title: string;
  /** @nullable */
  bloodTestId?: number | null;
}