# lib/api-zod/src/generated/types/markerHistory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Test Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface MarkerHistory {
  markerName: string;
  /** @nullable */
  testDate: string | null;
  /** @nullable */
  value: number | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  status: string | null;
  /** @nullable */
  referenceRangeLow: number | null;
  /** @nullable */
  referenceRangeHigh: number | null;
  bloodTestId: number;
}