# lib/api-zod/src/generated/types/uploadBloodTestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Test Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadBloodTestBody {
  fileName: string;
  pdfBase64: string;
  /** @nullable */
  notes?: string | null;
}